Extract reusable diagnostic state to separate object
